+++ src/app/diagram/calculation.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 
 import { CalculationService } from './calculation.service';
-import { ToRgbaPipe } from './shared/to-rgba.pipe';
+import { ToRgbaPipe } from '../shared/to-rgba.pipe';
 
 describe('CalculationService', () => {
     let service: CalculationService;
