+++ src/main.ts
@@ -3,8 +3,8 @@ import { AppComponent } from './app/app.component';
 import { bootstrapApplication } from '@angular/platform-browser';
 import { appRoutes } from './app/app.routes';
 import { importProvidersFrom } from '@angular/core';
-import { CalculationService } from './app/calculation.service';
-import { DataService } from './app/data.service';
+import { CalculationService } from './app/diagram/calculation.service';
+import { DataService } from './app/shared/data.service';
 
 import { enableProdMode, ɵprovideZonelessChangeDetection } from '@angular/core';
 import { environment } from './environments/environment';
